pyproject.toml
[project]
name = "github-explain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.8.0",
    "mcp[cli]>=1.6.0",
    "rich>=14.0.0",
    "typing-extensions>=4.13.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "D", "I"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"
indent-style = "space"
